feat : winner check and print function add
